* commit '80c867f685324dc64cded890f50fb84bfd601556': (217 commits)
  KERNEL-418 更新升级重构
  REPORT-22587 新版本插件管理，下载资源交互样式
  REPORT-19946 Linux设计器
  REPORT-19946 Linux设计器
  sonar问题修复
  REPORT-22291 撤销
  REPORT-20329 sonar问题修复
  REPORT-22291 模版版本控制，远程设计刚升级到10.0.4的工程，默认配置项开启
  REPORT-19946 Linux设计器
  KERNEL-418 更新升级重构
  KERNEL-418 更新升级重构
  MOBILE-23325 【单选按钮样式】图文按钮：上传图片后，移动端图片显示很模糊
  KERNEL-418 更新升级重构
  无jira任务,sonar问题修复
  无jira任务,sonar问题修复
  KERNEL-418 更新升级重构
  REPORT-19946 Linux设计器
  REPORT-19946 Linux设计器
  KERNEL-418 更新升级重构
  CHART-10233 插件超链
  ...

+++ designer-base/src/main/java/com/fr/design/mainframe/backgroundpane/TextureBackgroundQuickPane.java
@@ -101,7 +101,7 @@ public class TextureBackgroundQuickPane extends BackgroundQuickPane {
 			Dimension d = getSize();
 
 			g2d.setPaint(this.buttonTexturePaint);
-			GraphHelper.fill(g2d, new Rectangle2D.Double(0, 0, d.width - 1, d.height - 1));
+			GraphHelper.fill(g2d, new Rectangle2D.Double(0, 0, (double) d.width - 1, (double)d.height - 1));
 
 			if (ComparatorUtils.equals(texturePaint, this.buttonTexturePaint)) {// it's
 				// selected.
@@ -109,7 +109,7 @@ public class TextureBackgroundQuickPane extends BackgroundQuickPane {
 			} else {
 				g2d.setPaint(null);
 			}
-			GraphHelper.draw(g2d, new Rectangle2D.Double(0, 0, d.width - 1, d.height - 1));
+			GraphHelper.draw(g2d, new Rectangle2D.Double(0, 0, (double)d.width - 1, (double) d.height - 1));
 		}
 
 		public Dimension getPreferredSize() {
